Feature/final impl

+++ app/src/main/java/com/example/reflect/presentation/adapter/RecordsListAdapter.kt
@@ -21,8 +21,8 @@ import java.util.Calendar
 
 class RecordsListAdapter(
     private val calendar: Calendar,
-    private val onEdit: (Int, RecordModel) -> Unit,
-    private val onDelete: (Int) -> Unit
+    private val onEdit: ((Int, RecordModel) -> Unit)? = null,
+    private val onDelete: ((Int) -> Unit)? = null
 ) : ListAdapter<GetRecordsState, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
 
     class EmptyRecordsViewHolder(
@@ -37,7 +37,13 @@ class RecordsListAdapter(
         private val binding: CardStateBinding
     ) : RecyclerView.ViewHolder(binding.root) {
         @SuppressLint("SetTextI18n")
-        fun bind(model: RecordModel, today: Calendar, context: Context, onDelete: (Int) -> Unit, onEdit: (Int, RecordModel) -> Unit) {
+        fun bind(
+            model: RecordModel,
+            today: Calendar,
+            context: Context,
+            onDelete: ((Int) -> Unit)?,
+            onEdit: ((Int, RecordModel) -> Unit)?
+        ) {
             with (binding) {
                 val currentDate = Calendar.getInstance()
                 currentDate.time = model.creationDate!!
@@ -47,23 +53,23 @@ class RecordsListAdapter(
                 when(model.value) {
                     in 0..1 -> {
                         cardStateImageView.setImageResource(R.drawable.ic_state_image_1)
-                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, "Ужасное")
+                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, context.resources.getString(R.string.awfulState))
                     }
                     in 2..3 -> {
                         cardStateImageView.setImageResource(R.drawable.ic_state_image_2)
-                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, "Плохое")
+                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, context.resources.getString(R.string.badState))
                     }
                     in 4..6 -> {
                         cardStateImageView.setImageResource(R.drawable.ic_state_image_3)
-                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, "Нормально")
+                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, context.resources.getString(R.string.normalState))
                     }
                     in 7..8 -> {
                         cardStateImageView.setImageResource(R.drawable.ic_state_image_4)
-                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, "Хорошее")
+                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, context.resources.getString(R.string.goodState))
                     }
                     in 9..10 -> {
                         cardStateImageView.setImageResource(R.drawable.ic_state_image_5)
-                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, "Отличное")
+                        cardStateChangeMoodTV.text = context.resources.getString(R.string.cardStateMood, context.resources.getString(R.string.excellentState))
                     }
                     else -> throw IllegalStateException("Как так вообще получилось, что значение от 0 до 10 больше 10?!")
                 }
@@ -91,23 +97,27 @@ class RecordsListAdapter(
                     cardStateSecondRV.adapter = RecordsTagListAdapter(model.secondTagList)
                 }
 
-                cardStateChangeDots.setOnClickListener {
-                    val popupMenu = PopupMenu(context, cardStateChangeDots)
-                    popupMenu.inflate(R.menu.card_state_menu)
-                    popupMenu.setOnMenuItemClickListener {
-                        when(it.itemId) {
-                            R.id.menuEdit -> {
-                                onEdit(model.id, model)
-                                true
+                if (onEdit == null || onDelete == null) {
+                    cardStateChangeDots.visibility = View.GONE
+                } else {
+                    cardStateChangeDots.setOnClickListener {
+                        val popupMenu = PopupMenu(context, cardStateChangeDots)
+                        popupMenu.inflate(R.menu.card_state_menu)
+                        popupMenu.setOnMenuItemClickListener {
+                            when(it.itemId) {
+                                R.id.menuEdit -> {
+                                    onEdit(model.id, model)
+                                    true
+                                }
+                                R.id.menuDelete -> {
+                                    onDelete(model.id)
+                                    true
+                                }
+                                else -> false
                             }
-                            R.id.menuDelete -> {
-                                onDelete(model.id)
-                                true
-                            }
-                            else -> false
                         }
+                        popupMenu.show()
                     }
-                    popupMenu.show()
                 }
             }
         }
